Moved inline arraycall to NumPy module
