+++ numba/core/cpu.py
@@ -144,6 +144,7 @@ class CPUContext(BaseContext):
         # fix for #8940
         from numba.np.unsafe import ndarray # noqa F401
         # TODO: Hide all NumPy features behind a flag
+        from numba.np import inline_arraycall # noqa F401
 
     @property
     def target_data(self):
